@@ -3,10 +3,11 @@ const User = require("../models/users.model");
 const crypto = require("crypto");
 const KEY = process.env.KEY;
 const HASH_ALGORITHM_TO_USE = crypto.getHashes().find(itemName => itemName === "sha256");
-mongoose.set("useFindAndModify", false);
 
-exports.newUser = (req, res) => {
-  const passwordHash = crypto
+
+exports.newUser = async (req, res) => {
+  try {
+    const passwordHash = crypto
     .createHash(HASH_ALGORITHM_TO_USE, KEY)
     // updating data
     .update(req.body.passWord)
@@ -20,16 +21,22 @@ exports.newUser = (req, res) => {
     dateOfCreation: req.body.creationDate,
   });
 
-  newUserProfile.save((err, data) => {
-    if (err) {
-      res.status(400).send({
-        data: "",
-        ErrorMessage: `Error occured while adding user to database: user not added. Info: ${err}`,
-      });
-    } else {
-      res.status(200).send({ data: data._id });
-    }
-  });
+  let savedUserProfile = await newUserProfile.save();
+
+  if(savedUserProfile) {
+    res.status(200).send({ data: savedUserProfile._id });
+  } else {
+    res.status(400).send({
+      data: null,
+      ErrorMessage: `Error occured while adding user to database: user not added. Info: ${err}`,
+    });
+  }
+  } catch(error) {
+    res.status(400).send({
+      data: null,
+      ErrorMessage: `Error occured while adding user to database: user not added. Info: ${err}`,
+    });
+  }
 };
 
 exports.retrieveAllUsers = (req, res) => {
@@ -45,47 +52,56 @@ exports.retrieveAllUsers = (req, res) => {
   });
 };
 
-exports.authenticateUser = (req, res) => {
-  User.find({ userName: req.body.userID }, (err, data) => {
-    if (err) {
-      res.status(400).send({
+exports.authenticateUser = async (req, res) => {
+  try {
+    let response = await User.find({ userName: req.body.userID }).exec();
+      if (response.length == 0) {
+        console.log("wrong")
+        res.status(400).send({
+          authStatus: "false",
+          user: "",
+          ErrorMessage: "Error occured while retrieving records from database",
+        });
+      } else {
+        let passwordHash = crypto
+          .createHash(HASH_ALGORITHM_TO_USE, KEY)
+          // updating data
+          .update(req.body.passWord)
+          // Encoding to be used
+          .digest("hex");
+  
+        let userNameHash = crypto
+          .createHash(HASH_ALGORITHM_TO_USE, KEY)
+          // updating data
+          .update(req.body.userID)
+          // Encoding to be used
+          .digest("hex");
+  
+        let authStatusHash = crypto
+          .createHash(HASH_ALGORITHM_TO_USE, KEY)
+          // updating data
+          .update(userNameHash + process.env.SALT)
+          // Encoding to be used
+          .digest("hex");
+          
+        if (response[0].password === passwordHash) {
+          res.status(200).send({
+            authStatus: authStatusHash,
+            user: userNameHash + process.env.SALT,
+            ErrorMessage: "",
+          });
+        } else {
+          res
+            .status(200)
+            .send({ authStatus: "false", user: "", ErrorMessage: "" });
+        }
+      }
+  } catch(error) {
+    console.log(error)
+          res.status(400).send({
         authStatus: "false",
         user: "",
         ErrorMessage: "Error occured while retrieving records from database",
       });
-    } else {
-      let passwordHash = crypto
-        .createHash(HASH_ALGORITHM_TO_USE, KEY)
-        // updating data
-        .update(req.body.passWord)
-        // Encoding to be used
-        .digest("hex");
-
-      let userNameHash = crypto
-        .createHash(HASH_ALGORITHM_TO_USE, KEY)
-        // updating data
-        .update(req.body.userID)
-        // Encoding to be used
-        .digest("hex");
-
-      let authStatusHash = crypto
-        .createHash(HASH_ALGORITHM_TO_USE, KEY)
-        // updating data
-        .update(userNameHash + process.env.SALT)
-        // Encoding to be used
-        .digest("hex");
-        
-      if (data[0] && data[0].password === passwordHash) {
-        res.status(200).send({
-          authStatus: authStatusHash,
-          user: userNameHash + process.env.SALT,
-          ErrorMessage: "",
-        });
-      } else {
-        res
-          .status(200)
-          .send({ authStatus: "false", user: "", ErrorMessage: "" });
-      }
-    }
-  });
+  }
 };
